github-integration/requirements.txt
PyGithub
python-dotenv 